Fixed an issue with connecting two Origins simultaneously on one Edge. (#197) And organized the source codes related to Orchestrator.
