+++ src/projects/orchestrator/AMS.mk
@@ -3,6 +3,12 @@ include $(DEFAULT_VARIABLES)
 
 LOCAL_TARGET := orchestrator
 
+LOCAL_SOURCE_FILES := $(LOCAL_SOURCE_FILES) \
+    $(call get_sub_source_list,data_structures) \
+
+LOCAL_HEADER_FILES := $(LOCAL_HEADER_FILES) \
+    $(call get_sub_header_list,data_structures) \
+
 $(call add_pkg_config,openssl)
 $(call add_pkg_config,srt)
