+++ src/projects/orchestrator/data_structures/data_structure.h
@@ -1,0 +1,13 @@
+//==============================================================================
+//
+//  OvenMediaEngine
+//
+//  Created by Hyunjun Jang
+//  Copyright (c) 2019 AirenSoft. All rights reserved.
+//
+//==============================================================================
+#pragma once
+
+#include "enums.h"
+#include "interfaces.h"
+#include "structures.h"
