+++ src/projects/providers/rtmp/rtmp_provider.cpp
@@ -53,7 +53,7 @@ namespace pvd
 		auto server = GetServerConfig();
 		auto rtmp_address = ov::SocketAddress(server.GetIp(), static_cast<uint16_t>(server.GetBind().GetProviders().GetRtmp().GetPort().GetPort()));
 
-		_physical_port = PhysicalPortManager::Instance()->CreatePort(ov::SocketType::Tcp, rtmp_address);
+		_physical_port = PhysicalPortManager::GetInstance()->CreatePort(ov::SocketType::Tcp, rtmp_address);
 		if (_physical_port == nullptr)
 		{
 			logte("Could not initialize phyiscal port for RTMP server: %s", rtmp_address.ToString().CStr());
@@ -68,7 +68,7 @@ namespace pvd
 	bool RtmpProvider::Stop()
 	{
 		_physical_port->RemoveObserver(this);
-		PhysicalPortManager::Instance()->DeletePort(_physical_port);
+		PhysicalPortManager::GetInstance()->DeletePort(_physical_port);
 
 		_physical_port = nullptr;
