+++ src/projects/providers/rtsp/rtp/rtp_udp_track.h
@@ -80,7 +80,7 @@ public:
         for (const auto &port : port_range)
         {
             address.SetPort(port);
-            auto physical_port = PhysicalPortManager::Instance()->CreatePort(socket_type, address);
+            auto physical_port = PhysicalPortManager::GetInstance()->CreatePort(socket_type, address);
             if (physical_port)
             {
                 if (rtp_physical_port == nullptr)
